FEM: Make new analysis more user friendly
New FEM wb users quite often complain that thay can't add new constrains
after creating a new analysis. The reason is that a new analysis hides
shape and exposes newly created mesh, but constraints have to be added
to the shape, so the mesh has to be hidden first. This patch changes the
behaviour: newly created mesh is semi-transparent and non-selectable, so
user can add constraints without hiding the mesh and showing the shape.

Signed-off-by: Przemo Firszt <przemo@firszt.eu>